Fix #732, change uint32 to size_t

+++ src/os/rtems/src/os-impl-network.c
@@ -57,7 +57,7 @@ int32 OS_NetworkGetID_Impl(int32 *IdBuf)
  *           See prototype for argument/return detail
  *
  *-----------------------------------------------------------------*/
-int32 OS_NetworkGetHostName_Impl(char *host_name, uint32 name_len)
+int32 OS_NetworkGetHostName_Impl(char *host_name, size_t name_len)
 {
     int32 return_code;
